pyproject.toml
[project]
name = "koiz-data-portal"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.135.3",
    "paho-mqtt>=2.1.0",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.44.0",
]
